refactor: complete structure of html, fix some styles
