+++ src/assets/mock/newsList.ts
@@ -35,4 +35,32 @@ export const news: NewsItem[] = [
       текст.. Здесь ваш текст.."`,
     published: '11.12.2019',
   },
+  {
+    id: 2,
+    title: 'Второй заголовок лайфхака',
+    comments: 10,
+    likes: 21,
+    favorites: 7,
+    text: `Давно выяснено, что при оценке дизайна и композиции читаемый текст
+      мешает сосредоточиться. Lorem Ipsum используют потому, что тот
+      обеспечивает более или менее стандартное заполнение шаблона, а
+      также реальное распределение букв и пробелов в абзацах, которое не
+      получается при простой дубликации "Здесь ваш текст.. Здесь ваш
+      текст.. Здесь ваш текст.."`,
+    published: '11.12.2019',
+  },
+  {
+    id: 3,
+    title: 'Второй заголовок лайфхака',
+    comments: 10,
+    likes: 21,
+    favorites: 7,
+    text: `Давно выяснено, что при оценке дизайна и композиции читаемый текст
+      мешает сосредоточиться. Lorem Ipsum используют потому, что тот
+      обеспечивает более или менее стандартное заполнение шаблона, а
+      также реальное распределение букв и пробелов в абзацах, которое не
+      получается при простой дубликации "Здесь ваш текст.. Здесь ваш
+      текст.. Здесь ваш текст.."`,
+    published: '11.12.2019',
+  },
 ];
