+++ src/routes/PublicRoute/styled.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 
 export const Main = styled.main`
-  background-color: #fff;
+  /* background-color: #fff; */
   min-height: calc(100vh - 220px); /*minus header and footer height*/
 `;
